requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
pandas>=1.1.0
tqdm>=4.48.0
Pillow>=7.2.0
scikit-image>=0.17.0
opencv-python>=4.4.0
tensorboard>=2.3.0 